% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloo.R
\name{eloo-class}
\alias{eloo-class}
\alias{new_eloo}
\alias{eloo}
\title{\code{eloo} vector}
\usage{
new_eloo(value = double(), pareto_k = double())

eloo(df)
}
\arguments{
\item{value}{a numeric vector}

\item{pareto_k}{a numeric vector of Pareto-\emph{k} shape parameters corresponding
to \code{x}}

\item{df}{a data frame with columns \code{value} and \code{pareto_k}}
}
\value{
An S3 vector of class \code{eloo}. The Pareto \emph{k} parameters are
accessible using \code{\link[vctrs:fields]{vctrs::field()}}.
}
\description{
Attach Pareto-\emph{k} estimates to expectations from the \code{loo} package. Mostly
intended for internal development.
}
\details{
The \code{eloo} class makes it easier to work with vectors that have Pareto \emph{k}
values attached. For a conservative set of basic arithmetic operations,
\code{eloo} objects will retain their Pareto \emph{k} values, but most operations will
fall gracefully back to standard numeric vector operations.
}
\section{Functions}{
\itemize{
\item \code{new_eloo}: Create an \code{eloo} vector.

\item \code{eloo}: Create an \code{eloo} vector from the output of
\code{\link[loo:E_loo]{loo::E_loo()}}.
}}

