% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfit.R
\name{new_rdfit}
\alias{new_rdfit}
\title{Constructor for \code{rdfit}}
\usage{
new_rdfit(
  data,
  cohort_colname,
  group_colname,
  person_colname,
  testlet_colname,
  item_colname,
  max_score_colname,
  obs_score_colname
)
}
\arguments{
\item{data}{a data frame with columns for cohort, group, person, testlet,
item, maximum score, and observed score}

\item{cohort_colname}{name of the cohort column}

\item{group_colname}{name of the group column}

\item{person_colname}{name of the person column}

\item{testlet_colname}{name of the testlet column}

\item{item_colname}{name of the item column}

\item{max_score_colname}{name of the maximum score column}

\item{obs_score_colname}{name of the observed score column}
}
\value{
An \code{rdfit} object.
}
\description{
Construct an \code{rdfit} object.
}
