% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfit.R
\name{rdfit-class}
\alias{rdfit-class}
\alias{rdfit}
\alias{new_rdfit}
\title{Fit a \code{raschdash} model}
\usage{
rdfit(
  data,
  cohort = "All",
  group = "All",
  person,
  testlet = "All",
  item,
  max_score,
  obs_score,
  k = 1,
  ...
)

new_rdfit(
  cohorts,
  groups,
  persons,
  person_groups,
  testlets,
  items,
  item_testlets,
  item_max_scores,
  obs_group_cohorts,
  obs_groups,
  obs_group_items,
  obs_group_scores,
  obs_person_cohorts,
  obs_persons,
  obs_person_items,
  obs_person_scores,
  k = 1,
  pars = c("standard", "logit", "hyper", "all"),
  ...
)
}
\arguments{
\item{data}{a data frame with columns for cohort, group, person, testlet,
item, maximum possible item score, and observed score. For
group-level observations, the person should be \code{NA}.}

\item{cohort}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> cohort column. Defaults to a
generic \emph{All} cohort.}

\item{group}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> group column. Defaults to a
generic \emph{All} group.}

\item{person}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> person column}

\item{testlet}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> testlet column. Defaults to a
generic \emph{All} testlet.}

\item{item}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> item column}

\item{max_score}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> maximum-score column}

\item{obs_score}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> observed score column}

\item{k}{maximum score of rating-scale items. Defaults to no rating scales.}

\item{...}{additional parameters passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{cohorts}{vector of unique cohort identifiers}

\item{groups}{vector of unique group identifiers}

\item{persons, person_groups}{vectors of unique person identifiers (\code{persons})
and the identifiers of their corresponding
groups (\code{person_groups})}

\item{testlets}{vector of unique testlet identifiers}

\item{items, item_testlets, item_max_scores}{vectors of unique item identifiers
(\code{items}) and their corresponding
testlets (\code{item_testlets}) and
maximum scores(\code{max_scores})}

\item{obs_group_cohorts, obs_groups, obs_group_items, obs_group_scores}{for the group-level observations, vectors of the corresponding
cohorts (\code{obs_group_cohorts}), groups (\code{obs_groups}), items
(\code{obs_group_items}), and observed scores (\code{obs_group_scores})}

\item{obs_person_cohorts, obs_persons, obs_person_items, obs_person_scores}{for the person-level observations, vectors of the corresponding
cohorts (\code{obs_person_cohorts}), persons (\code{obs_persons}), items
(\code{obs_person_items}), and observed scores (\code{obs_person_scores})}

\item{pars}{which parameters to save from the Stan model. Defaults to
\code{standard}, which only saves parameters on the standard scale,
but may also be \code{logit} to add logit-scale parameters,
or \code{hyper} to add logit-scale parameters and hyper-parameters.
For debugging, the \code{all} option adds other raw parameters
specific to the Stan implementation.}
}
\value{
An object of class \code{rdfit} with the following attributes.
\describe{
\item{\code{data}}{The input data as passed to Stan.}
\item{\code{stanfit}}{
The \link[rstan:stanfit-class]{rstan::stanfit} object for the fitted model. It contains samples of
the following parameters.
\itemize{
\item standard-scale parameters
\itemize{
\item \code{group_ability}: group abilities
\item \code{person_ability}: person abilities
\item \code{testlet_difficulty}: testlet difficulties
\item \code{item_difficulty}: item difficulties
\item \code{thresholds}: rating-scale threshold offsets from overall item
difficulty (if any rating-scale items are
present)
\item \code{sigma}: standard deviation of person abilities on the logit scale.
Used to convert between the logit and standard scales.
}
\item posterior predictive distribution
\itemize{
\item \code{y_rep}: predicted score for each observed combination of
group, person, and item
\item \code{prior_group_ability} (sampled under posterior hyper-parameters)
\item \code{prior_person_ability} (sampled under posterior hyper-parameters)
\item \code{prior_testlet_difficulty} (sampled under posterior hyper-parameters)
\item \code{prior_item_difficulty} (sampled under posterior hyper-parameters)
}
\item log likelihoods
\itemize{
\item \code{log_lik}: log likelihood of observed scores
\item \code{log_lik_rep}: log likelihood of predicted scores
}
\item logit-scale parameters (when \code{pars = "logit"} or \code{pars = "hyper"})
\itemize{
\item \code{xi}: group abilities
\item \code{eta}: person abilities
\item \code{epsilon}: testlet difficulties
\item \code{delta}: item difficulties
\item \code{tau}: rating-scale threshold offsets from overall item
difficulty (if any rating-scale items are present)
}
\item hyper-parameters (when \code{pars = "hyper"})
\itemize{
\item \code{nu}: mean difficulty
\item \code{psi}: scale of group abilities
\item \code{phi}: scale of person abilities (relative to group)
\item \code{theta_epsilon}: scale of testlet difficulties
\item \code{theta_upsilon}: scale of item difficulties (relative to testlet)
\item \code{theta}: scale of item difficulties (overall)
}
}
}

\item{\code{loo}}{
A \link[loo:loo]{loo::loo} object for \code{stanfit}.
}
}
}
\description{
Create \code{rdfit} objects from observed data.
}
\details{
The \code{rdfit()} function is intended to be the primary method for users to create
\code{rdfit} objects. Most R users work with denormalised \emph{dimensional data} in a
single data frame. In the case that the data are already stored in a
normalised structure, however, the raw \code{new_rdfit()} constructor will work.
The \code{new_rdfit()} constructor does not validate whether the data are properly
normalised, however, and Stan will throw sometimes cryptic error messages if
they are not.
}
\section{Functions}{
\itemize{
\item \code{rdfit}: Create an \code{rdfit} object from dimensional data.

\item \code{new_rdfit}: Create an \code{rdfit} object from normalised data.
}}

