% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschdash-package.R
\name{raschdash-package}
\alias{raschdash-package}
\title{\code{raschdash}: Dashboards for Rasch models in the classroom}
\description{
\lifecycle{experimental}
The \code{raschdash} package provides two tools.
\enumerate{
\item Fully Bayesian Rasch models with calibrations and diagnostics for
binary, binomial, and rating-scale items.
\item Templates for visualising the calibrations and diagnostics from these
models, including space to add comments per cohort.
}
}
