% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschdash-package.R
\docType{package}
\name{raschdash-package}
\alias{raschdash}
\alias{raschdash-package}
\title{raschdash: Rasch-Model Dashboards for Classroom Testing}
\description{
Dashboard generator for exploring calibrations and diagnostics
    from Bayesian Rasch-model fits, intended to support classroom testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jaburgoyne/raschdash}
  \item Report bugs at \url{https://github.com/jaburgoyne/raschdash/issues}
}

}
\author{
\strong{Maintainer}: John Ashley Burgoyne \email{j.a.burgoyne@uva.nl}

}
\keyword{internal}
