% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfit.R
\name{rdfit-class}
\alias{rdfit-class}
\alias{new_rdfit}
\alias{rdfit}
\title{Fit a \code{raschdash} model}
\usage{
new_rdfit(
  cohorts,
  groups,
  persons,
  person_groups,
  testlets,
  items,
  item_testlets,
  item_max_scores,
  obs_group_cohorts,
  obs_groups,
  obs_group_items,
  obs_group_scores,
  obs_person_cohorts,
  obs_persons,
  obs_person_items,
  obs_person_scores,
  k = 0,
  ...
)

rdfit(
  data,
  cohort = NULL,
  group,
  person,
  testlet,
  item,
  max_score,
  obs_score,
  k = 0,
  ...
)
}
\arguments{
\item{cohorts}{vector of unique cohort identifiers}

\item{groups}{vector of unique group identifiers}

\item{persons, person_groups}{vectors of unique person identifiers (\code{persons})
and the identifiers of their corresponding
groups (\code{person_groups})}

\item{testlets}{vector of unique testlet identifiers}

\item{items, item_testlets, item_max_scores}{vectors of unique item identifiers
(\code{items}) and their corresponding
testlets (\code{item_testlets}) and
maximum scores(\code{max_scores})}

\item{obs_group_cohorts, obs_groups, obs_group_items, obs_group_scores}{for the group-level observations, vectors of the corresponding
cohorts (\code{obs_group_cohorts}), groups (\code{obs_groups}), items
(\code{obs_group_items}), and observed scores (\code{obs_group_scores})}

\item{obs_person_cohorts, obs_persons, obs_person_items, obs_person_scores}{for the person-level observations, vectors of the corresponding
cohorts (\code{obs_person_cohorts}), persons (\code{obs_persons}), items
(\code{obs_person_items}), and observed scores (\code{obs_person_scores})}

\item{k}{maximum score of rating-scale items. Defaults to no rating scales.}

\item{...}{additional parameters passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{data}{a data frame with columns for cohort, group, person, testlet,
item, maximum possible item score, and observed score. For
group-level observations, the person should be \code{NA}.}

\item{cohort}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> cohort column. Defaults to a
generic \emph{All} cohort.}

\item{group}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> group column}

\item{person}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> person column}

\item{testlet}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> testlet column}

\item{item}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> item column}

\item{max_score}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> maximum-score column}

\item{obs_score}{<\link[tidyr:tidyr_tidy_select]{tidyr::tidyr_tidy_select}> observed score column}
}
\value{
An \code{rdfit} object, which has the following attributes.
\describe{
\item{stanfit}{the \link[rstan:stanfit-class]{rstan::stanfit} object for the fitted
model. It contains samples of the following parameters.
\itemize{
\item logit-scale parameters
\describe{
\item{xi}{group abilities}
\item{eta}{person abilities}
\item{epsilon}{testlet difficulties}
\item{delta}{item difficulties}
\item{tau}{rating-scale threshold offsets from
overall item difficulty (if any
rating-scale items are present)}
\item{lambda}{standard deviation of person
abilities. Used to transform the
logit scale to the standard scale.}
}
\item standard-scale parameters
\describe{
\item{group_ability}{group abilities}
\item{person_ability}{person abilities}
\item{testlet_difficulty}{testlet difficulties}
\item{item_difficulty}{item difficulties}
\item{thresholds}{rating-scale threshold offsets
from overall item difficulty (if
any rating-scale items are
present)}
}
\item standard-scale parameters under their prior
distribution given the posterior distribution of
the hyper-parameters
\describe{
\item{prior_group_ability}{}
\item{prior_person_ability}{}
\item{prior_testlet_difficulty}{}
\item{prior_item_difficulty}{}
}
\item posterior predictive distribution
\describe{
\item{y_rep}{predicted score for each observed
combination of group, person, and item}
}
\item log likelihoods
\describe{
\item{log_lik}{log likelihood of observed scores}
\item{log_lik_rep}{log likelihood of predicted
scores}
}
}}
\item{loo}{a \link[loo:loo]{loo::loo} object for \code{stanfit}}
}
}
\description{
Create \code{rdfit} objects from observed data.
}
\details{
The \code{rdfit} function is intended to be the primary method for users to create
\code{rdfit} objects. Most R users work with so-called \emph{dimensional data} in a
single data frame. In the case that the data are alread stored in a
normalised form, however, the raw \code{new_rdfit} constructor will work. Be
aware that \code{new_rdfit} does not validate whether the data are properly
normalised, and Stan will throw rather cryptic error message if they are not.
}
\section{Functions}{
\itemize{
\item \code{new_rdfit}: Create an \code{rdfit} object from normalised data.

\item \code{rdfit}: Create an \code{rdfit} object from dimensional data.
}}

